[ALPS07340433] imgsys: fix coverity
1. fix coverity for secure patch

MTK-Commit-Id: a7a76f87aba80ff24bb144774de17a14a7469031

CR-Id: ALPS07340433
Feature: [Android Default] Camera Application Basic Functions

+++ drivers/media/platform/mtk-isp/imgsys/mtk_imgsys-sys.c
@@ -1451,7 +1451,7 @@ static void imgsys_scp_handler(void *data, unsigned int len, void *priv)
 		return;
 	}
 
-	if ((swbuf_data->offset < 0) ||
+	if (((int)(swbuf_data->offset) < 0) ||
 		(swbuf_data->offset > mtk_hcp_get_gce_mem_size(imgsys_dev->scp_pdev))) {
 		pr_info("%s: invalid swbuf_data->offset(%d), max(%d)\n",
 			__func__, swbuf_data->offset,
